Update Makefile.EEE
Even if EEE (ESS EPICS Environment) is slowly phased out,
it is good to have a Makefike with all source files inside this Git
to allow other people to write Makefiles for their make system.
Update the list of source files.
Change ethercatlib to be "e3", working together with the latest
EtherCAT master from Etherlab, see

https://github.com/icshwi/etherlabmaster

  commit 442b7ad526b7c93e9ce7220eef85a9e48bc14ca0
  Author: Jeong Han Lee <jeonghan.lee@gmail.com>
  Date:   Sun Nov 11 01:10:23 2018 +0100