Update  expression end char to "|" (before was "#")
This is only valid for the "Cfg.SetPLCExpr(...)" and "Cfg.AppendPLCExpr(...)" commands.

Background:
Preferably a ";" should be used in the end of each expression according to ExprTK syntax.
Unfortenately the ";" sign is used as command delimiter in the asynOctet interface where
several commands can be stacked in the same command string. Therefore an alternative char
have been used to end exprTk expressions.
In the "Cfg.LoadPlcFile(...)" the normal exprTk syntax with ";" should be used.