Add optional ec slave revision number to Cfg.EcSlaveVerify() cmd.
If a revsion number, other than 0, is specified in Cfg.EcSlaveVerify(), a comparison between the revsion numbers
of the config vs the actual slave will be performed.
The revsion number of the slave needs to be higher or equal compared to the config value otherwise
Cfg.EcSlaveVerify() will return a error.

This functionality is to avoid complications that could accur if config revison is newer than actual revsion of
the slave on the bus since the slave might not support the functions defined in the config (or addresses changed).