ecmcTrajectoryS.cpp
When using ruckig s-curev the follwing restrictions apply for when in pos mode:
On the fly change of target velocity is not allowed. A new target velo will be used at next start of next positioning command. On the fly change of target position is allowed
when in velo mode:
On the fly change of target velocity is allowed.

Note:
When using trapetz trajectory, the velocity target can be on the fly update in both pos and velo mode. Target pos can also be updated on the fly when in pos mode.